+++ src/test/java/com/ing/data/cassandra/jdbc/UsingDseContainerTest.java
@@ -15,7 +15,7 @@ package com.ing.data.cassandra.jdbc;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.AfterAll;
-import org.testcontainers.containers.CassandraContainer;
+import org.testcontainers.cassandra.CassandraContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
@@ -30,13 +30,13 @@ abstract class UsingDseContainerTest {
     // Using @Container annotation restarts a new container for each test of the class, so as it takes ~20/30 sec. to
     // start a Cassandra container, we just want to have one container instance for all the tests of the class. See:
     // https://www.testcontainers.org/test_framework_integration/manual_lifecycle_control/#singleton-containers
-    static CassandraContainer<?> cassandraContainer;
+    static CassandraContainer cassandraContainer;
 
     protected static void initializeContainer() {
         // For the official DataStax Enterprise server image, see here: https://hub.docker.com/r/datastax/dse-server/
         final DockerImageName dseServerImage = DockerImageName.parse("datastax/dse-server:7.0.0-a")
             .asCompatibleSubstituteFor("cassandra");
-        cassandraContainer = new CassandraContainer<>(dseServerImage)
+        cassandraContainer = new CassandraContainer(dseServerImage)
             .withEnv("DS_LICENSE", "accept")
             .withEnv("CLUSTER_NAME", "embedded_test_cluster")
             .withEnv("DC", "datacenter1")